changed the code for good practice
