+++ lib/presentation/model/movie_item.dart
@@ -6,19 +6,23 @@ import 'package:newfluttermovieapp/domain/model/movie_domain.dart';
 
 class MovieItem {
   final int movieId;
-  String posterImg;
-  String title;
-  int vote;
-  String overview;
-  String releaseDate;
+ final String posterImg;
+ final String title;
+ final String overview;
+ final String releaseDate;
+  final double popularity;
+  final String rating;
+  final int voteCount;
 
   MovieItem({
     @required this.movieId,
     this.posterImg,
     this.title,
-    this.vote,
     this.overview,
     this.releaseDate,
+    this.rating,
+    this.popularity,
+    this.voteCount
   });
 
 }
