--- lib/presentation/movie_list/params/movie_list_param.dart
@@ -1,7 +1,0 @@
-// import 'package:newfluttermovieapp/domain/model/movie_domain.dart';
-//
-// class MovieListParams{
-//  final List<MovieDomain> movieList ;
-//  MovieListParams(this.movieList);
-//
-// }
